api/requirements.txt
fastapi==0.115.12
pandas==2.2.3
pydantic==2.11.2
sentence_transformers==3.3.1
torch==2.4.1
uvicorn==0.34.0
# This is not listed as an explicit import but is necessary for the SentenceTransformer
huggingface-hub==0.25.1
azure-storage-blob==12.16.0
